+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- New axes provider for allowing axes to be updated.
 - Bug fixes and updates.
 - Tooltips refined.
 - New client layout. Tooltips now uses a store instead of being a provider.